Initial commit
A simple but working parsing of the mbr records from the block device
with an integration test running with a bash script to create a simple
fat32 fs, which could be used also for future tests.

This is more like a warm up, as I provided the initial definition of
the errors and so on.