Refactoring and more testing
When adding binrw, I've did a bit of a mess with the error and result
and conversions. This is a partial cleanup. I've removed the Read/Write
binrw's io traits for the moment, I've readded only Write because it was
used in a test.

I've added a new feature gate for std. The plan is to add support to std
via a feature flag, and eventually add support to the standard io
traits.

More tests are added for the timesamp. According to intellij the
coverage is around 85%, but it should be even a bit higher.

I've also found lack of a couple of tests and added few more todos here
and there.