Refactor mutex inside cached partition
and remove impl for the blockdevice trait from here. The CachedParition
is used as middleware to interface with the underneath blockdevice, and
I'm always going to use it (because I will need functions like _offset).
Also there is information related to the physical structure of the
device that is needed when reading and writing. I can keep this
information in the cachedpartition which should help simplifying the
cluster reader/writer structs.

Currently I need a fully ARC clone in order to access the device, but
maybe I don't need it. Actually, maybe I don't need ARC at all. Will
need to figure it out.