Refactor directory.contents() and fixes a small bug along the way
previously, if last cluster couldn't accomodate the required spots, I
would allocate a new cluster and use that cluster starting with offset
zero.
In this way, I was putting empty DirectoryEntries at the end of the
previous cluster, end these are equivalent to "EndOfEntries".

Previous logic to update_entry_by_index, was reading till the end of the
allocated clusters, and filtering out EndOfEntries entries. The
assumption broken here was that EndOfEntries padding was at the end not
in between. Thanks to the filtering, the previous logic was able to
update the right index, and appear like working while IRL it wasn't.

I'm planning to further simplify the logic, and just use CCW/CCR
everywhere I can.